requirements.txt
# Quant Trading System - Core Dependencies
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0

# MT5 Bridge - ZeroMQ (optional, only needed for ZMQ bridge mode)
pyzmq>=25.0.0

# Technical Analysis
ta>=0.10.0

# Regime-Switching Strategy (Instruct.md)
yfinance>=0.2.0
matplotlib>=3.7.0
scipy>=1.10.0
scikit-optimize>=0.9.0
pytz>=2023.3

